Adding a little project.
While surfing the web, I saw a button which let one share their typing speed; I took a look at the URL and played around with it. I was immediately compelled to make a program which would allow one to type in their desired typing speed and get it. Don't cheat if you just wanted to share this with your friends. Just thought it would be cool to see how the URL works.